requirements.txt
pandas
black[jupyter]
plotly
pydantic
pydantic[email]
mysql-connector-python
ipdb
python-dotenv
faker
